All tests Window Feature Manager ok.
